extraction-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
python-multipart>=0.0.6
pillow>=10.2.0
pymupdf>=1.23.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
structlog>=24.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.2.0
black>=24.1.0
mypy>=1.8.0
tenacity>=8.2.3
openai>=1.10.0
anthropic>=0.18.0
lxml>=5.1.0
pika
